DBZ-2972 Handle IP only for user to avoid false matches;
If for example a label is used without space like `label1:BEGIN`, then the
tokenizer would create tokens `label1`, `:BE` and `GIN` as `:BE` matches
`IP_ADDRESS` token. As the tokenization is globally executed it is
better to skip a separate token and link IPv6/IPv4 directly to the
username token.